chore: refactor home and oven device
